+++ tests/CacheTest.php
@@ -18,8 +18,6 @@ class CacheTest extends TestCase
 {
 
     /** @test
-     * @throws \Exception
-     * @throws \GuzzleHttp\Exception\GuzzleException
      */
     public function requests_are_cached() : void
     {
@@ -39,13 +37,11 @@ class CacheTest extends TestCase
         $responseTwo = $client->request('GET', '/')->getProtocolVersion();
         $responseThree = $client->request('GET', '/')->getProtocolVersion();
 
-        $this->assertNotEquals($responseOne, $responseTwo);
-        $this->assertEquals($responseTwo, $responseThree);
+        static::assertNotEquals($responseOne, $responseTwo);
+        static::assertEquals($responseTwo, $responseThree);
     }
 
     /** @test
-     * @throws \Exception
-     * @throws \GuzzleHttp\Exception\GuzzleException
      */
     public function throw_too_many_requests_when_nothing_in_cache() : void
     {